Refactor `bevy_render_2d::material`
