+++ crates/bevy_render_2d/src/material/specialization.rs
